Refactor branding page into tabbed sub-pages
