+++ resources/views/partials/branding-header.blade.php
@@ -1,7 +1,7 @@
 <div class="relative mb-6 w-full">
-    <flux:heading size="xl" level="1">{{ __('Settings') }}</flux:heading>
+    <flux:heading size="xl" level="1">{{ __('Branding') }}</flux:heading>
     <flux:subheading size="lg" class="mb-6">
-        {{ __('Manage your profile and account settings') }}
+        {{ __('Customize your studio branding and visual identity.') }}
     </flux:subheading>
     <flux:separator variant="subtle" />
 </div>
